+++ src/cmd/watch.rs
@@ -22,9 +22,9 @@ impl Watch {
     pub async fn run(self, config: Option<PathBuf>) -> Result<()> {
         let (shutdown_tx, _shutdown_rx) = broadcast::channel(1);
         let cfg = ConfigOpts::rtc_watch(self.build, self.watch, config)?;
-        let mut system = WatchSystem::new(cfg, shutdown_tx.clone(), None).await?;
+        let system = WatchSystem::new(cfg, shutdown_tx.clone(), None).await?;
 
-        system.build().await.ok();
+        system.trigger_build();
         let system_handle = tokio::spawn(system.run());
         tokio::signal::ctrl_c()
             .await