+++ src/serve.rs
@@ -57,9 +57,14 @@ impl ServeSystem {
 
     /// Run the serve system.
     #[tracing::instrument(level = "trace", skip(self))]
-    pub async fn run(mut self) -> Result<()> {
+    pub async fn run(self) -> Result<()> {
+        // Perform an initial build
+        let mut build_done_rx = self.build_done_chan.subscribe();
+        self.watch.trigger_build();
+        let _build_res = build_done_rx.recv().await; // TODO: only open after a successful build.
+        drop(build_done_rx);
+
         // Spawn the watcher & the server.
-        let _build_res = self.watch.build().await; // TODO: only open after a successful build.
         let watch_handle = tokio::spawn(self.watch.run());
         let server_handle = Self::spawn_server(
             self.cfg.clone(),